test: move index_reorg_crash to baseindex_tests
It tests generic functionality, nothing specific to the
blockfilterindex.

Mostly move-only, can be reviewed with
--color-moved=dimmed-zebra --color-moved-ws=ignore-all-space
